Refactor and enhance `updateWhere` methods in ObjectCollection and JsonCollection
The commit solves a typo on a parameter name, changing "filedSelector" to "fieldSelector". It also enhances the `updateWhere` methods in ObjectCollection and JsonCollection classes to support more data types for the `fieldValue` parameter. Additionally, new `updateWhere` methods are introduced, which handle encoding to JSON element internally for ease of use.